Crate hyperlocal

hyperlocal provides hyper client and server bindings for unix domain sockets

See the hyperlocal::UnixConnector docs for how to configure hyper clients and the hyperlocal::server::Http docs for how to configure hyper servers

Re-exports

pub use client::UnixConnector;

Modules

client
Hyper client bindings for unix domain sockets

server
Hyper server bindings for unix domain sockets

Structs

Uri
A type which implements Into for hyper's hyper::Uri type targetting unix domain sockets.