Struct hyperlocal::server::Http

pub struct Http<B = Chunk> { /* fields omitted */ }

A type that provides a factory interface for creating unix socket based Servers

examples

extern crate hyper;
extern crate hyperlocal;

//let server = hyperlocal::Http::new().bind(
 // "path/to/socket",
 //  || Ok(HelloWorld)
//)

Methods

impl<B: AsRef<[u8]> + 'static> Http<B>

fn new() -> Http<B>

Creates a new instance of the HTTP protocol, ready to spawn a server or start accepting connections.

fn bind<P, S, Bd>(&self, path: P, new_service: S) -> Result<Server<S, Bd>> where
P: AsRef<Path>,
S: NewService<Request = Request, Response = Response<Bd>, Error = Error> + Send + Sync + 'static,
Bd: Stream<Item = B, Error = Error> + 'static,

binds a new server instance to a unix domain socket path