Struct hyperlocal::UnixSocketConnector

pub struct UnixSocketConnector;

A type which implements hyper's NetworkConnector trait for unix domain sockets A type which implements hyper's IntoUrl interface for unix domain sockets. You can use this with any of the HTTP factory methods on hyper's Client interface.

examples

 extern crate hyper;
 extern crate hyperlocal;

 let client = hyper::Client::with_connector(
     hyperlocal::UnixSocketConnector
 );