Crate hyperlocal

hyperlocal provides hyper client and server bindings for unix domain sockets

See the UnixSocketConnector docs for how to configure hyper clients and the UnixSocketServer docs for how to configure hyper servers

Structs

DomainUrl
A type which implements hyper's IntoUrl interface for unix domain sockets. You can use this with any of the HTTP factory methods on hyper's Client interface.

UnixSocketConnector
A type which implements hyper's NetworkConnector trait for unix domain sockets A type which implements hyper's IntoUrl interface for unix domain sockets. You can use this with any of the HTTP factory methods on hyper's Client interface.

UnixSocketListener
A type which implements hyper's NetworkListener trait

UnixSocketServer
A type that provides a factory interface for creating unix socket based hyper Servers

UnixSocketStream
A type which implements hyper's NetworkStream trait