Struct hyperlocal::server::Http

pub struct Http { /* fields omitted */ }

A lower-level method of creating a unix domain socket server.

This structure is used to manage connections yourself. If you don't need to do this, consider using the higher-level Server::bind API.

Examples

extern crate hyper;
extern crate hyperlocal;

use std::os::unix::net::UnixListener;
use hyper::{Response, rt::{Future, Stream}, service::service_fn};
use hyperlocal::server::{Http, Incoming};

let listener = UnixListener::bind("hyperlocal_test_echo_server_2.sock").unwrap();
let incoming = Incoming::from_std(listener, &Default::default()).unwrap();
let serve = Http::new().serve_incoming(
  incoming,
  move || service_fn(
    |req| Ok::<_, hyper::Error>(Response::new(req.into_body()))
  )
 );

let server = serve.for_each(|connecting| {
    connecting
    .then(|connection| {
        let connection = connection.unwrap();
        Ok::<_, hyper::Error>(connection)
    })
    .flatten()
    .map_err(|err| {
        std::io::Error::new(
            std::io::ErrorKind::Other,
            format!("failed to serve connection: {}", err),
        )
    })
});

Methods

impl Http

pub fn new() -> Self

Creates a new instance of the HTTP protocol, ready to spawn a server or start accepting connections.

pub fn from_hyper(hyper_http: HyperHttp) -> Self

Creates a new instance of the HTTP protocol using the given hyper Http, ready to spawn a server or start accepting connections.

pub fn serve_path<P, S>(&self, path: P, new_service: S) -> Result<Serve<S>> where
P: AsRef<Path>,
S: NewService<ReqBody = Body>,
S::ResBody: Payload,
S::Service: Send,
S::Error: Into<Box<dyn Error + Send + Sync>>,
<S::Service as Service>::Future: Send + 'static,

Bind the provided path with the default Handle and return Serve.

This method will bind the unix domain socket path provided with a new UDS listener ready to accept connections. Each connection will be processed with the new_service object provided, creating a new service per connection.

If the provided path already exists, this method will return an error.

pub fn serve_path_handle<P, S>(
&self,
path: P,
handle: &Handle,
new_service: S
) -> Result<Serve<S>> where
P: AsRef<Path>,
S: NewService<ReqBody = Body> + Send + 'static,
S::ResBody: Payload,
S::Service: Send,
S::Error: Into<Box<dyn Error + Send + Sync>>,
<S::Service as Service>::Future: Send + 'static,

Bind the provided path with the Handle and return Serve.

This method will bind the unix domain socket path provided with a new UDS listener ready to accept connections. Each connection will be processed with the new_service object provided, creating a new service per connection.

If the provided path already exists, this method will return an error.

pub fn serve_incoming<S>(&self, incoming: Incoming, new_service: S) -> Serve<S> where
S: NewService<ReqBody = Body>,
S::ResBody: Payload,
S::Error: Into<Box<dyn Error + Send + Sync>>,

Bind the provided stream of incoming UnixStream objects with a NewService.